+++ src/main/java/org/ohdsi/webapi/analysis/converter/AnalysisConceptSetToConceptSetConverter.java
@@ -3,7 +3,9 @@ package org.ohdsi.webapi.analysis.converter;
 import org.ohdsi.webapi.analysis.AnalysisConceptSet;
 import org.ohdsi.webapi.converter.BaseConversionServiceAwareConverter;
 import org.ohdsi.webapi.service.dto.ConceptSetDTO;
+import org.springframework.stereotype.Component;
 
+@Component
 public class AnalysisConceptSetToConceptSetConverter<T extends AnalysisConceptSet> extends BaseConversionServiceAwareConverter<T, ConceptSetDTO> {
  
     @Override
