+++ src/main/java/org/ohdsi/webapi/common/DesignImportService.java
@@ -14,21 +14,41 @@ import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
+import org.ohdsi.webapi.analysis.AnalysisConceptSet;
+import org.ohdsi.webapi.conceptset.ConceptSetItem;
+import org.ohdsi.webapi.service.ConceptSetService;
+import org.ohdsi.webapi.service.dto.ConceptSetDTO;
+import org.springframework.context.annotation.DependsOn;
+import org.springframework.core.convert.ConversionService;
 
 @Service
 public class DesignImportService {
-
     private final Security security;
     private final UserRepository userRepository;
     private final CohortDefinitionRepository cohortRepository;
     private final CohortDefinitionDetailsRepository detailsRepository;
+    private final ConversionService conversionService;
+    private final ConceptSetService conceptSetService;
 
-    public DesignImportService(Security security, UserRepository userRepository, CohortDefinitionRepository cohortRepository, CohortDefinitionDetailsRepository detailsRepository) {
-
+    public DesignImportService(Security security, UserRepository userRepository, CohortDefinitionRepository cohortRepository, CohortDefinitionDetailsRepository detailsRepository, ConceptSetService conceptSetService, ConversionService conversionService) {
         this.security = security;
         this.userRepository = userRepository;
         this.cohortRepository = cohortRepository;
         this.detailsRepository = detailsRepository;
+        this.conceptSetService = conceptSetService;
+        this.conversionService = conversionService;
+    }
+    
+    public ConceptSetDTO persistConceptSet(final AnalysisConceptSet analysisConceptSet) {
+        ConceptSetDTO cs = conversionService.convert(analysisConceptSet, ConceptSetDTO.class);
+        cs = conceptSetService.createConceptSet(cs);
+        final Integer conceptSetId = cs.getId();
+        List<ConceptSetItem> csi = Arrays.stream(analysisConceptSet.expression.items).map(i -> conversionService.convert(i, ConceptSetItem.class)).collect(Collectors.toList());
+        csi.forEach(n -> n.setConceptSetId(conceptSetId));
+        conceptSetService.saveConceptSetItems(cs.getId(), csi.stream().toArray(ConceptSetItem[]::new));
+        return cs;
     }
 
     public CohortDefinition persistCohortOrGetExisting(final CohortDefinition cohort) {
@@ -51,21 +71,20 @@ public class DesignImportService {
     }
 
     private Optional<CohortDefinition> findCohortByExpressionHashcode(final CohortDefinitionDetails details) {
-        List<CohortDefinitionDetails> detailsFromDb = detailsRepository.findByHashCode(details.calculateHashCode());
-        return detailsFromDb
-                .stream()
-                .filter(v -> Objects.equals(v.getStandardizedExpression(), details.getStandardizedExpression()))
-                .findFirst()
-                .map(CohortDefinitionDetails::getCohortDefinition);
+        return this.findCohortByExpressionHashcodeAndPredicate(details, (c -> true));
     }
     
     private Optional<CohortDefinition> findCohortByExpressionHashcodeAndName(final CohortDefinitionDetails details, final String cohortName) {
+        return this.findCohortByExpressionHashcodeAndPredicate(details, c -> Objects.equals(c.getName(), cohortName));
+    }
+    
+    private Optional<CohortDefinition> findCohortByExpressionHashcodeAndPredicate(final CohortDefinitionDetails details, final Predicate<CohortDefinition> c) {
         List<CohortDefinitionDetails> detailsFromDb = detailsRepository.findByHashCode(details.calculateHashCode());
         return detailsFromDb
                 .stream()
                 .filter(v -> Objects.equals(v.getStandardizedExpression(), details.getStandardizedExpression()))
                 .map(CohortDefinitionDetails::getCohortDefinition)
-                .filter(c -> c.getName().equals(cohortName))
+                .filter(c)
                 .findFirst();
     }
 }
